Adding missing elements (without children)
I was playing around with gomponents as I like this approach and found
that HR and BR tags are not yet supported. As they need to have attributes
and there is no difference for attributes and children elements I added
them but made them into an extra file. I wonder if attributes and child
elements should not be handled as separate parameters.

My testis are testing a bit more than just the plain elements. Not sure if
you like that. Having a bigger scope for the tests makes me more
confident in having a good test coverage of the actual usage. But then
it may fail for code outside of the tested scope. What do you think
about that?

Currently there is also no way to enforce HTML rules like having LI
inside UL or OL and other stuff. I think it would be cool if this would
be enforced by the type checker or as an "strict" mode at runtime.

But I find it a very nice idea as is already. I dislike to add iraw HTML
code inside the source and I also dislike to mix HTML (or template) code
with dynamic code like if/then/else or loops.