Add Fragment
This change adds a `Fragment` top-level function, to render multiple elements without a parent element.

It's basically the same as `Group`, except the function signature matches the style of the other functions by being variadic.

But I also made `Group` renderable directly, with the important caveat that root-level attributes are _ignored_.
I don't think this will give problems in practice, as the main use case for `Fragment` is basically to return
root-level elements to the client using something like HTMX.

Fixes #162