Remove SVG package
Because the added value of having this was very small, and I don't want to confuse people about what's properly supported,
I've removed the SVG package. As #196 showed, there can easily be conflicts between HTML and SVG, and the rendering isn't
really set up for anything else than HTML.

It's still easy to create SVG elements: `html.SVG` exists in this repo, and the rest is often just a string that can be included using `g.Raw`.

Fixes #206