Move module to maragu.dev/gomponents namespace (#215)
This is a breaking change to move gomponents to my own import namespace.

I will obviously be careful with this, test it out in all kinds of
scenarios, release betas, etc. But otherwise, because the `Node`
interface is so simple and has basically never changed, I don't think
this will break much. 🤞

Fixes #200