Add html2gomponents tool
This is a separate tool for converting HTML to gomponents.
I decided to have a go at this after all, like @amrojjeh got started in #178,
but with the HTML parser built by the Go team. I actually like that there's an
authorative converter tool.

This lives in an entirely separate module in the same repo. I'm not yet
sure what the implications are of this, but I'm going to give it a try.

See #167